+++ CHANGELOG.md
@@ -178,6 +178,7 @@ here is a non-exhaustive list of significant departures from the original gem:
   This matches the behavior of DSL rules.
   Note that this only works for {OpenHAB::DSL::Rules::Terse terse rules} if they're created within a {OpenHAB::DSL::Rules::Builder rules.build} block.
 * {OpenHAB::DSL::Rules::BuilderDSL#on_start #on_start} supports creating a `core.SystemStartlevelTrigger`. Also see {OpenHAB::DSL::Rules::BuilderDSL#on_load #on_load}
+* {OpenHAB::DSL::Rules::Terse.calculated_item} (also available from {OpenHAB::DSL::Items::ItemBuilder#state})
 
 ### Bug Fixes